% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.pense.R
\name{coef.pensem}
\alias{coef.pensem}
\title{Extract Model Coefficients}
\usage{
\method{coef}{pensem}(object, sparse = FALSE, ...)
}
\arguments{
\item{object}{object of type \code{pensem} to extract coefficients from.}

\item{sparse}{return a sparse vector or a dense (base R) numeric vector}

\item{...}{currently not used.}
}
\value{
A numeric vector of size \eqn{p + 1}.
}
\description{
Extract Model Coefficients
}
