% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{consistency_const}
\alias{consistency_const}
\title{Get the Constant for Consistency for the M-Scale}
\usage{
consistency_const(delta, rho)
}
\arguments{
\item{delta}{desired breakdown point (between 0 and 0.5)}

\item{rho}{the name of the chosen \eqn{\rho} function.}
}
\value{
consistency constant
}
\description{
Get the Constant for Consistency for the M-Scale
}
\seealso{
Other miscellaneous functions: 
\code{\link{rho_function}()}
}
\concept{miscellaneous functions}
