% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-internal.R
\name{.approx_match}
\alias{.approx_match}
\title{Approximate Value Matching}
\usage{
.approx_match(x, table, eps)
}
\arguments{
\item{x, table}{see \link[base:match]{base::match} for details.}

\item{eps}{numerical tolerance for matching.}
}
\value{
a vector the same length as \code{x} with integers giving the position in
\code{table} of the first match if there is a match, or \code{NA_integer_}
otherwise.
}
\description{
Approximate Value Matching
}
\keyword{internal}
