% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-internal.R
\name{.bisquare_consistency_const}
\alias{.bisquare_consistency_const}
\title{Get the Constant for Consistency for the M-Scale Using the Bisquare Rho Function}
\usage{
.bisquare_consistency_const(delta)
}
\arguments{
\item{delta}{desired breakdown point (between 0 and 0.5)}
}
\value{
consistency constant
}
\description{
Get the Constant for Consistency for the M-Scale Using the Bisquare Rho Function
}
\keyword{internal}
