% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-internal.R
\name{.mopt_consistency_const}
\alias{.mopt_consistency_const}
\title{Get the Constant for Consistency for the M-Scale Using the Optimal Rho Function}
\usage{
.mopt_consistency_const(delta, eps = sqrt(.Machine$double.eps))
}
\arguments{
\item{delta}{desired breakdown point (between 0 and 0.5)}

\item{eps}{numerical tolerance for equality comparisons}
}
\value{
consistency constant
}
\description{
Get the Constant for Consistency for the M-Scale Using the Optimal Rho Function
}
\keyword{internal}
