% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-internal.R
\name{.mopt_efficiency_const}
\alias{.mopt_efficiency_const}
\title{Get the constant for the desired efficiency of the M-estimate of location
using the optimal \eqn{\rho} function}
\usage{
.mopt_efficiency_const(eff, eps = sqrt(.Machine$double.eps))
}
\arguments{
\item{eff}{desired efficiency (between 0 and 1)}

\item{eps}{numerical tolerance for equality comparisons}
}
\value{
tuning constant for desired efficiency
}
\description{
Get the constant for the desired efficiency of the M-estimate of location
using the optimal \eqn{\rho} function
}
\keyword{internal}
