% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-internal.R
\name{.run_replicated_cv}
\alias{.run_replicated_cv}
\title{Run replicated K-fold CV with random splits}
\usage{
.run_replicated_cv(
  std_data,
  cv_k,
  cv_repl,
  cv_est_fun,
  metric,
  par_cluster = NULL,
  handler_args = list()
)
}
\arguments{
\item{std_data}{standardized full data set
(standardized by \code{.standardize_data})}

\item{cv_k}{number of folds per CV split}

\item{cv_repl}{number of CV replications.}

\item{cv_est_fun}{function taking the standardized training set and
the indices of the left-out observations and returns a list of estimates.
The function always needs to return the same number of estimates!}

\item{metric}{function taking a vector of prediction errors and
returning the scale of the prediction error.}

\item{par_cluster}{parallel cluster to parallelize computations.}

\item{handler_args}{additional arguments to the handler function.}
}
\description{
Run replicated K-fold CV with random splits
}
\keyword{internal}
