% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-internal.R
\name{.standardize_data}
\alias{.standardize_data}
\title{Standardize data}
\usage{
.standardize_data(
  x,
  y,
  intercept,
  standardize,
  robust,
  sparse,
  mscale_opts,
  location_rho = "bisquare",
  cc,
  target_scale_x = NULL,
  ...
)
}
\arguments{
\item{x}{predictor matrix. Can also be a list with components \code{x} and \code{y},
in which case \code{y} is ignored.}

\item{y}{response vector.}

\item{intercept}{is an intercept included (i.e., should \code{y} be centered?)}

\item{standardize}{standardize or not.}

\item{robust}{use robust standardization.}

\item{location_rho}{rho function for location estimate}

\item{cc}{cutoff value for the rho functions used in scale and location
estimates.}

\item{...}{passed on to \code{mlocscale()}.}
}
\value{
a list with the following entries:
}
\description{
Standardize data
}
\keyword{internal}
