% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/elnet.R
\name{elnet}
\alias{elnet}
\title{Elastic Net Regression}
\usage{
elnet(X, y, alpha, lambda, maxit = 10000, eps, centering = TRUE,
  addLeading1s = TRUE, en.algorithm = c("augmented-lars",
  "coordinate-descent", "augmented-lars-gram", "augmented-lars-nogram"))
}
\arguments{
\item{X}{The data matrix X}

\item{y}{The response vector}

\item{alpha,lambda}{The values for the parameters controlling the penalization}

\item{maxit}{The maximum number of iterations}

\item{eps}{The relative tolerance for convergence for gradient-descent (default 1e-8) or
the threshold for treating numbers as 0 in LARS (default .Machine$double.eps)}

\item{centering}{Should the rows be centered first}

\item{addLeading1s}{Should a leading column of 1's be appended? If \code{FALSE}, this has
to be done before calling this function.}

\item{en.algorithm}{algorithm to use to compute the elastic net solution.}
}
\value{
\item{coefficients}{The regression coefficients}
        \item{residuals}{The residuals}
        \item{converged}{Did the algorithm converge?}
}
\description{
Compute the elastic net regression coefficients
}
\details{
This solves the minimization problem
\deqn{\frac{1}{2 N} RSS + \lambda \left( (1 - \alpha) \frac{1}{2} \| \beta \|_1 + \alpha \| \beta \|_2^2 \right)}{
     (1/2N) RSS + \lambda ( (1 - \alpha) (1/2) L1(\beta) + \alpha L2(\beta)^2 )}
}

