% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet.R
\name{elnet_cv}
\alias{elnet_cv}
\title{Cross-validate Elastic Net}
\usage{
elnet_cv(X, y, alpha, lambda, weights, intercept = TRUE, cv_k = 10,
  cv_measure, ncores = getOption("mc.cores", 1L), cl = NULL, options, ...)
}
\arguments{
\item{X}{data matrix with predictors}

\item{y}{response vector}

\item{alpha, lambda}{values for the parameters controlling the penalization}

\item{weights}{an optional vector of weights to be used in the fitting
process. Should be \code{NULL} or a numeric vector. If non-NULL,
weighted EN is used with weights \code{weights}. See also 'Details'.}

\item{intercept}{should an intercept be estimated?}

\item{cv_k}{number of cross validation folds.}

\item{cv_measure}{function (name) which takes a matrix of residuals and
returns a performance measure for each column.}

\item{ncores, cl}{the number of processor cores or an actual \code{parallel}
cluster. At most \code{cv_k + 1} nodes will be used.}

\item{...}{additional arguments passed on to \code{\link{elnet}}.}
}
\value{
\item{lambda}{vector of lambda values.}
 \item{status}{integer specifying the exit status of the EN algorithm.}
 \item{message}{explenation of the exit status.}
 \item{coefficients}{matrix of regression coefficients. Each
     column corresponds to the estimate for the lambda value at the same
     index.}
 \item{residuals}{matrix of residuals. Each column corresponds to the
     residuals for the lambda value at the same index.}
 \item{cvres}{data frame with lambda, average cross-validated performance
     and the estimated standard deviation.}
}
\description{
Perform k-fold cross-validation for \code{\link{elnet}}.
}
