% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{en_admm_options}
\alias{en_admm_options}
\title{Use the ADMM Elastic Net Algorithm}
\usage{
en_admm_options(max_it = 1000, step_size, acceleration = 1)
}
\arguments{
\item{max_it}{maximum number of iterations.}

\item{step_size}{step size for the algorithm.}

\item{acceleration}{acceleration factor for linearized ADMM.}
}
\value{
options for the ADMM EN algorithm.
}
\description{
Use the ADMM Elastic Net Algorithm
}
\seealso{
Other EN algorithms: 
\code{\link{en_dal_options}()},
\code{\link{en_lars_options}()}
}
\concept{EN algorithms}
