% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{en_ridge_options}
\alias{en_ridge_options}
\title{Ridge optimizer using an Augmented data matrix.
Only available for Ridge problems (`alpha=0``) and selected automatically in this case.}
\usage{
en_ridge_options()
}
\description{
Ridge optimizer using an Augmented data matrix.
Only available for Ridge problems (`alpha=0``) and selected automatically in this case.
}
\keyword{internal}
