% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initest.control.R
\name{enpy.control}
\alias{enpy.control}
\title{Control Parameters for ENPY}
\usage{
enpy.control(enMaxIt = 50000, enEPS = 1e-08, enCentering = TRUE,
  mscaleMaxIt = 200, mscaleEPS = 1e-08, mscaleRhoFun = c("bisquare",
  "huber", "gauss"))
}
\arguments{
\item{enMaxIt}{Maximum number of iterations allowed for the elastic net algorithm.}

\item{enEPS}{Convergence threshold for elastic net.}

\item{enCentering}{Should the data be centered for the elastic net algorithm. Only those rows
with leading 1's will be considered for centering.}

\item{mscaleMaxIt}{Maximum number of iterations allowed for the m-scale algorithm.}

\item{mscaleEPS}{Convergence threshold for the m-scale}

\item{mscaleRhoFun}{The rho function to use for the m-scale.}
}
\value{
A list with the given arguments.
}
\description{
Set control parameters for ENPY.
}

