% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpy.exact.R
\name{enpy.exact}
\alias{enpy.exact}
\title{PY (Pena-Yohai) initial estimates for EN}
\usage{
enpy.exact(X, y, alpha, lambda, deltaesc, cc.scale, psc.method = c("Mn",
  "Qp"), prosac, clean.method = c("threshold", "proportion"), C.res, prop,
  py.nit, en.tol, control)
}
\arguments{
\item{X}{The data matrix X -- a leading column of 1's will be added!}

\item{y}{The response vector.}

\item{lambda, alpha}{The EN penalty parameters (NOT adjusted for the number of observations
in \code{X}).}

\item{deltaesc, cc.scale}{Parameters for the M-equation of the scale. The default
rho function is Tukey's bisquare. This can be changed by the parameter \code{control}.}

\item{psc.method}{The method to use for computing the principal sensitivity components.
See details.}

\item{prosac}{The proportion of observations to remove based on PSCs.}

\item{clean.method}{How to clean the data based on large residuals.
If \code{"threshold"}, all observations with scaled residuals larger than
\code{C.res} will be removed.
If \code{"proportion"}, observations with the largest \code{prop} residuals
will be removed.}

\item{C.res, prop}{How many or which observations should be removed based on their residuals.}

\item{py.nit}{The maximum number of iterations to perform.}

\item{en.tol}{The relative tolerance for convergence.}

\item{control}{Further control parameters (see \code{\link{enpy.control}})}
}
\value{
\item{coeff}{A numeric matrix with one initial coefficient per column}
        \item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
Computes the PY initial estimates for EN with exact principal sensitivity components
}
