% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enpy.rr.R
\name{enpy.rr}
\alias{enpy.rr}
\title{PY (Peña-Yohai) initial estimates for EN}
\usage{
enpy.rr(X, y, lambda1, lambda2, deltasc, cc.scale, prosac,
  clean.method = c("threshold", "proportion"), C.res, prop, py.nit, en.tol,
  en.centering = TRUE)
}
\arguments{
\item{X}{The data matrix X, with leading column of 1's}

\item{y}{The response vector}

\item{lambda1,lambda2}{The EN penalty parameters (adjusted for the number of observations
in \code{X})}

\item{deltasc,cc.scale}{Parameters for the M-equation of the scale. Tukey's bisquare rho function
is used internally.}

\item{prosac}{The proportion of observations to remove based on PSCs}

\item{clean.method}{How to clean the data based on large residuals.
If \code{"threshold"}, all observations with scaled residuals larger than
\code{C.res} will be removed.
If \code{"proportion"}, observations with the largest \code{prop} residuals
will be removed.}

\item{C.res,prop}{How many or which observations should be removed based on their residuals.}

\item{py.nit}{The maximum number of iterations to perform.}

\item{en.tol}{The relative tolerance for convergence.}

\item{en.centering}{Should rows with a leading 1 be centered in the elastic net algorithm.
     Default's to \code{TRUE}.}
}
\value{
\item{coeff}{A numeric matrix with one initial coefficient per column}
        \item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
Computes the PY initial estimates for EN with approximated principal sensitivity components
by the ridge regression solution.
}

