% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpy.R
\name{enpy_ols}
\alias{enpy_ols}
\title{PY (Pena-Yohai) initial estimates for EN S-Estimators}
\usage{
enpy_ols(X, y, options)
}
\arguments{
\item{X}{data matrix with predictors -- a leading column of 1's will be
added!}

\item{y}{response vector.}

\item{options}{additional options for the initial estimator. See
\code{\link{initest_options}} for details.}
}
\value{
\item{coeff}{A numeric matrix with one initial coefficient per column}
        \item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
Computes the PY initial estimates for EN S-Estimators with
principal sensitivity components approximated by the ridge regression
solution.
}
