% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpy.R
\name{enpy_rr}
\alias{enpy_rr}
\title{PY (Pena-Yohai) initial estimates for EN S-Estimators}
\usage{
enpy_rr(X, y, alpha, lambda, options, en_options)
}
\arguments{
\item{X}{data matrix with predictors -- a leading column of 1's will be
added!}

\item{y}{response vector.}

\item{lambda, alpha}{The EN penalty parameters (NOT adjusted for the number
of observations in \code{X}).}

\item{options}{additional options for the initial estimator. See
\code{\link{initest_options}} for details.}

\item{en_options}{additional options for the EN algorithm. See
\code{\link{en_options}} for details.}
}
\value{
\item{coeff}{A numeric matrix with one initial coefficient per column}
        \item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
Computes the PY initial estimates for EN S-Estimators with
principal sensitivity components approximated by the ridge regression
solution.
}
