% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.R
\name{initest.control}
\alias{initest.control}
\title{Creates the internal control list for PY initial estimators}
\usage{
initest.control(lambda, alpha, numIt, eps = 1e-06,
  resid.clean.method = c("proportion", "threshold"), resid.threshold,
  resid.proportion, psc.proportion, mscale.delta = 0.5, mscale.cc = 1.54764,
  enpy.control)
}
\arguments{
\item{lambda}{numeric >= 0}

\item{alpha}{numeric 0 <= alpha <= 1}

\item{numIt}{integer > 0}

\item{eps}{numeric > 0}

\item{resid.clean.method}{character}

\item{resid.threshold}{If \code{resid.clean.method = "threshold"} numeric > 0, otherwise not
referenced}

\item{resid.proportion}{If \code{resid.clean.method = "proportion"} numeric > 0, otherwise not
referenced}

\item{psc.proportion}{numeric > 0}

\item{mscale.delta}{numeric > 0}

\item{mscale.cc}{numeric > 0}
}
\description{
Takes care of the correct storage mode for the arguments passed to the C/C++ code
}

