% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{initest_options}
\alias{initest_options}
\title{Deprecated Options for Initial Estimates}
\usage{
initest_options(
  keep_solutions = 5,
  psc_method = c("exact", "rr"),
  maxit = 10,
  maxit_pense_refinement = 5,
  eps = 1e-06,
  psc_keep = 0.5,
  resid_keep_method = c("proportion", "threshold"),
  resid_keep_prop = 0.6,
  resid_keep_thresh = 2,
  mscale_eps = 1e-08,
  mscale_maxit = 200
)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge('deprecated')}
}
\details{
Superseded by \code{\link[=enpy_options]{enpy_options()}}.
}
\keyword{internal}
