% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{mstep}
\alias{mstep}
\title{Perform an M-step after the EN S-Estimator}
\usage{
mstep(penseobj, nlambda = 100L, lambda, cv_k = 5L, scale,
  ncores = getOption("mc.cores", 1L), cl = NULL,
  options = mstep_options(), X, y, cv_objective, en_options, standardize)
}
\arguments{
\item{penseobj}{an object returned from a call to \code{\link{pense}}.}

\item{nlambda}{if \code{lambda} is not given,
a grid of \code{nlambda} lambda values is generated based on the data.}

\item{lambda}{a single value or a grid of values for the regularization
parameter of the M-step.
Assumed to be on the same scale as the data.
If missing, a grid of lambda
values is automatically generated (see parameter \code{nlambda}).
If supplied and \code{standardize = TRUE}, the lambda values will be
adjusted accordingly.}

\item{cv_k}{perform k-fold CV to choose the optimal lambda for prediction.}

\item{scale}{initial scale estimate for the M step. By default the
S-scale from the initial estimator (\code{penseobj}) is used.}

\item{ncores, cl}{use multiple cores or the supplied cluster for the
cross-validation. See \code{\link{pense}} for more details.}

\item{options}{additional options for the M-step.}

\item{X, y, cv_objective, en_options, standardize}{override arguments
provided to the original call to \code{\link{pense}}.}

\item{lambda}{regularization parameter for the MM-estimator.
If missing, a grid of lambda values is chosen automatically.}
}
\description{
Performs an M-step using the S-estimator at the optimal penalty
parameter as returned from \code{\link{pense}} as the initial
estimate.
}
