% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{mstep_options}
\alias{mstep_options}
\title{Deprecated Additional Options for the Penalized EN MM-estimator}
\usage{
mstep_options(
  cc = 3.44,
  maxit = 1000,
  eps = 1e-06,
  adjust_bdp = FALSE,
  verbosity = 0,
  en_correction = TRUE
)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge('deprecated')}
}
\details{
Superseded by \code{\link[=mm_algorithm_options]{mm_algorithm_options()}} and options supplied directly to \code{\link[=pense]{pense()}}.
}
\keyword{internal}
