% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pense.coldwarm.R
\name{pense.coldwarm}
\alias{pense.coldwarm}
\title{PENSE}
\usage{
\method{pense}{coldwarm}(X, y, alpha, lambda.grid, standardize, control)
}
\arguments{
\item{lambda.grid}{A grid of lambda values NOT ADJUSTED for sample size!}
}
\description{
PENSE
}

