% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initest.control.R
\name{pense.control}
\alias{pense.control}
\title{Control Parameters for PENSE}
\usage{
\method{pense}{control}(pense.maxit = 500, pense.tol = 1e-06,
  pense.en.tol = 1e-08, pense.en.maxit = 5e+05, init.tol = 1e-06,
  init.maxit = 10, init.resid.clean.method = c("proportion", "threshold"),
  init.resid.proportion = 0.4, init.resid.threshold = 2,
  init.psc.method = c("rr", "Mn"), init.psc.proportion = 0.8,
  init.csteps = 10, init.nkeep = 5, init.en.tol = pense.en.tol,
  init.en.maxit = pense.en.maxit, mscale.cc = 1.54764, mscale.delta = 0.5,
  mscale.maxit = 200, mscale.tol = 1e-08, mscale.rho.fun = c("bisquare",
  "huber", "gauss"))
}
\value{
A list with the given arguments.
}
\description{
Set control parameters for PENSE.
}

