% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{pense_options}
\alias{pense_options}
\title{Deprecated Additional Options for the Penalized EN S-estimator}
\usage{
pense_options(
  delta = 0.25,
  maxit = 1000,
  eps = 1e-06,
  mscale_eps = 1e-08,
  mscale_maxit = 200,
  verbosity = 0,
  cc = NULL,
  en_correction = TRUE
)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge('deprecated')}
}
\details{
Superseded by \code{\link[=mm_algorithm_options]{mm_algorithm_options()}} and options supplied directly to \code{\link[=pense]{pense()}}.
}
\keyword{internal}
