% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.cv_elnetfit}
\alias{plot.cv_elnetfit}
\title{Plot Method for Cross-Validated Elastic Net Models}
\usage{
\method{plot}{cv_elnetfit}(x, what = c("cv", "coef.path"), ...)
}
\arguments{
\item{x}{a fitted, cross-validated EN model from \code{\link{elnet_cv}}.}

\item{what}{what Plot either ther cross-validated prediction error
(\code{"cv"}; default) or the coefficient paths.}

\item{...}{currently ignored.}
}
\description{
Plot the cross-validation error or the coefficient path for a fitted
elastic net regression model.
}
