% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.elnetfit}
\alias{plot.elnetfit}
\title{Plot Method for Fitted Elastic Net Models}
\usage{
\method{plot}{elnetfit}(x, ...)
}
\arguments{
\item{x}{a fitted EN model from \code{\link{elnet}}.}

\item{...}{currently ignored.}
}
\description{
Plot the coefficient path for a fitted elastic net regression model.
}
