% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.pense}
\alias{plot.pense}
\title{Plot Method for Fitted Penalized Elastic Net S/MM-Estimates of Regression}
\usage{
\method{plot}{pense}(x, what = c("cv", "coef.path"), ...)
}
\arguments{
\item{x}{a PENSE estimate from \code{\link{pense}} or  \code{\link{pensem}}.}

\item{what}{what Plot either the cross-validated prediction error
(\code{"cv"}; default) or the coefficient paths.}

\item{...}{currently ignored.}
}
\description{
Plot the cross-validation error or the coefficient path for a fitted
PENSE estimate.
}
