% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.pense.R
\name{predict.pense}
\alias{predict.pense}
\title{Predict Method for Penalized Elastic-Net S-estimator}
\usage{
\method{predict}{pense}(object, newdata, lambda, ...)
}
\arguments{
\item{object}{an object of type \code{pense} to use for prediction.}

\item{newdata}{an optional design matrix}

\item{lambda}{the value of the penalty parameter. Default is to use the optimal
lambda \code{lambda.opt}.}

\item{...}{currently ignored.}
}
\value{
a numeric vector of predicted values for the given lambda.
}
\description{
Predict Method for Penalized Elastic-Net S-estimator
}

