% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prinsens.R
\name{prinsens}
\alias{prinsens}
\title{Principal Sensitivity Components}
\usage{
prinsens(X, y, intercept = TRUE)
}
\arguments{
\item{X}{The data matrix X}

\item{y}{The response vector}

\item{intercept}{Should an intercept be added or not.}
}
\value{
A numeric matrix with as many rows as \code{X} and as many columns as
     PSCs found (at most the number of columns in \code{X} plus one for the intercept).
     Each column is a PSC.
}
\description{
Compute the principal sensitiviy components (PSC) for regression.
}
\references{
Pena, D., & Yohai, V.. (1999). A Fast Procedure for Outlier Diagnostics in Large
Regression Problems. \emph{Journal of the American Statistical Association}, 94(446),
434–445. \url{http://doi.org/10.2307/2670164}
}

