% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.nsoptim_metrics}
\alias{print.nsoptim_metrics}
\title{Print Metrics}
\usage{
\method{print}{nsoptim_metrics}(x, max_level = NA, ...)
}
\arguments{
\item{x}{metrics object for printing.}

\item{max_level}{maximum level of printing which is applied for printing nested metrics.}
}
\description{
Pretty-print a list of metrics from optimization algorithm (if \code{pense} was built with metrics enabled).
}
\keyword{internal}
