% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pyinit.R
\name{pyinit}
\alias{pyinit}
\title{PY (Peña-Yohai) initial estimates for S-estimates of regression}
\usage{
pyinit(X, y, intercept = TRUE, deltaesc, cc.scale, prosac,
  clean.method = c("threshold", "proportion"), C.res, prop, py.nit, en.tol,
  control = enpy.control())
}
\arguments{
\item{X}{the data matrix X.}

\item{y}{the response vector.}

\item{intercept}{should an intercept be included in the models. Defaults to \code{TRUE}.}

\item{deltaesc,cc.scale}{parameters for the M-equation of the scale. Tukey's bisquare rho function
is used internally.}

\item{prosac}{the proportion of observations to remove based on PSCs.}

\item{clean.method}{how to clean the data based on large residuals.
If \code{"threshold"}, all observations with scaled residuals larger than
\code{C.res} will be removed (\code{C.res} corresponds to the constant
\eqn{C_1} from equation (21) in Pena & Yohai (1999).
If \code{"proportion"}, observations with the largest \code{prop} residuals
will be removed.}

\item{C.res,prop}{see parameter \code{clean.method} for details.}

\item{py.nit}{the maximum number of iterations to perform.}

\item{en.tol}{the relative tolerance for convergence.}

\item{control}{optional further control parameters from \code{\link{enpy.control}}.}
}
\value{
\item{initCoef}{A numeric matrix with one initial coefficient per column}
        \item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
Computes the PY initial estimates for S-estimates of regression.
}
\references{
Pena, D., & Yohai, V.. (1999). A Fast Procedure for Outlier Diagnostics in Large
Regression Problems. \emph{Journal of the American Statistical Association}, 94(446),
434–445. \url{http://doi.org/10.2307/2670164}
}

