% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals-methods.R
\name{residuals.elnetfit}
\alias{residuals.elnetfit}
\title{Extract Residuals from a Fitted Elastic-Net Estimator}
\usage{
\method{residuals}{elnetfit}(object, lambda, exact = FALSE,
  correction = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{elnet}} estimate to extract the residuals from.}

\item{lambda}{the value of the penalty parameter. Default is to use the
optimal lambda \code{object$lambda_opt} if the given estimator was
cross-validated or the smallest lambda if not.}

\item{exact}{if the lambda is not part of the lambda grid, should the
estimates be obtained by linear interpolation between the nearest
lambda values (default) or computed exactly.}

\item{correction}{should a correction factor be applied to the EN estimate?
See \code{\link{elnet}} for details on the applied correction.}

\item{...}{currently ignored.}
}
\value{
a numeric vector of residuals for the given lambda.
}
\description{
Extract Residuals from a Fitted Elastic-Net Estimator
}
