% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.pense.R
\name{residuals.pense}
\alias{residuals.pense}
\title{Extract Residuals from a Fitted Penalized Elastic-Net S/MM-estimator}
\usage{
\method{residuals}{pense}(object, lambda, exact = FALSE, correction = TRUE,
  ...)
}
\arguments{
\item{object}{a PENSE or PENSEM estimate to extract the residuals from.}

\item{lambda}{the value of the penalty parameter. Default is to use the
optimal lambda \code{object$lambda_opt}.}

\item{exact}{if the lambda is not part of the lambda grid, should the
estimates be obtained by linear interpolation between the nearest
lambda values (default) or computed exactly.}

\item{correction}{should a correction factor be applied to the EN estimate?
See \code{\link{elnet}} for details on the applied correction.}

\item{...}{currently ignored.}
}
\value{
a numeric vector of residuals for the given lambda.
}
\description{
Extract Residuals from a Fitted Penalized Elastic-Net S/MM-estimator
}
