% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setupCluster.R
\name{setupCluster}
\alias{setupCluster}
\title{Convenience function for parallel computing}
\usage{
setupCluster(ncores = 1L, cl = NULL, eval, export, envir = parent.frame())
}
\description{
This function handles setting up, using, and closing a potential cluster.
If no cluster of computing nodes is requested, it will create a proxy to the local
\code{lapply} function.
}

