% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tau_size}
\alias{tau_size}
\title{Compute the Tau-Scale of Centered Values}
\usage{
tau_size(x)
}
\arguments{
\item{x}{numeric values. Missing values are verbosely ignored.}
}
\value{
the \eqn{\tau} estimate of scale of centered values.
}
\description{
Compute the \eqn{\tau}-scale without centering the values.
}
\seealso{
Other functions to compute robust estimates of location and scale: 
\code{\link{mlocscale}()},
\code{\link{mloc}()},
\code{\link{mscale}()}
}
\concept{functions to compute robust estimates of location and scale}
