% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.pense.R
\name{coef.pense}
\alias{coef.pense}
\title{Extract Model Coefficients}
\usage{
\method{coef}{pense}(object, lambda, ...)
}
\arguments{
\item{object}{an object of type \code{pense} to extract coefficients from.}

\item{lambda}{the value of the penalty parameter. Default is to use the optimal
lambda \code{lambda.opt}.}

\item{...}{currently ignored.}
}
\value{
A numeric vector of size \eqn{p + 1}.
}
\description{
Extract Model Coefficients
}

