% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.pense.R
\name{coef.pense}
\alias{coef.pense}
\title{Extract Model Coefficients}
\usage{
\method{coef}{pense}(object, lambda, exact = FALSE, ...)
}
\arguments{
\item{object}{object of type \code{pense} to extract coefficients from.}

\item{lambda}{the value of the penalty parameter. Default is to use the
optimal lambda (\code{object$lambda_opt}).}

\item{exact}{if the lambda is not part of the lambda grid, should the
estimates be obtained by linear interpolation between the nearest
lambda values (default) or computed exactly.}

\item{...}{currently not used.}
}
\value{
A numeric vector of size \eqn{p + 1}.
}
\description{
Extract Model Coefficients
}
