% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{deprecated_en_options}
\alias{deprecated_en_options}
\alias{en_options_aug_lars}
\alias{en_options_dal}
\title{Deprecated Options for the EN Algorithms}
\usage{
en_options_aug_lars(use_gram = c("auto", "yes", "no"), eps = 1e-12)

en_options_dal(
  maxit = 100,
  eps = 1e-08,
  eta_mult = 2,
  eta_start_numerator = 0.01,
  eta_start,
  preconditioner = c("approx", "none", "diagonal"),
  verbosity = 0
)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge('deprecated')}
}
\section{Functions}{
\itemize{
\item \code{en_options_aug_lars}: Superseded by \code{\link[=en_lars_options]{en_lars_options()}}.

\item \code{en_options_dal}: Superseded by \code{\link[=en_dal_options]{en_dal_options()}}
}}

\seealso{
Other deprecated functions: 
\code{\link{enpy}()},
\code{\link{initest_options}()},
\code{\link{mstep_options}()},
\code{\link{pense_options}()},
\code{\link{pensem}()}
}
\concept{deprecated functions}
