% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{deprecated_en_options}
\alias{deprecated_en_options}
\alias{en_options_aug_lars}
\alias{en_options_dal}
\title{Deprecated}
\usage{
en_options_aug_lars(use_gram = c("auto", "yes", "no"), eps = 1e-12)

en_options_dal(
  maxit = 100,
  eps = 1e-08,
  eta_mult = 2,
  eta_start_numerator = 0.01,
  eta_start,
  preconditioner = c("approx", "none", "diagonal"),
  verbosity = 0
)
}
\arguments{
\item{use_gram}{\strong{ignored.} Should the Gram matrix be pre-computed.}

\item{eps}{\strong{ignored.} Numeric tolerance for convergence.}

\item{maxit}{maximum number of iterations allowed.}

\item{eta_mult}{multiplier to increase eta at each iteration.}

\item{eta_start_numerator}{if \code{eta_start} is missing, it is defined
by \code{eta_start = eta_start_numerator / lambda}.}

\item{eta_start}{\strong{ignored.} The start value for eta.}

\item{preconditioner}{\strong{ignored.} Preconditioner for the numerical solver. If none,
a standard solver will be used, otherwise the faster preconditioned
conjugate gradient is used.}

\item{verbosity}{\strong{ignored.}}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge('deprecated')}

Options for computing EN estimates.
}
\section{Functions}{
\itemize{
\item \code{en_options_aug_lars}: Superseded by \code{\link[=en_lars_options]{en_lars_options()}}.

\item \code{en_options_dal}: Superseded by \code{\link[=en_dal_options]{en_dal_options()}}
}}

\section{Warning}{
Do not use these functions in new code.
They may be removed from future versions of the package.
}

\seealso{
Other deprecated functions: 
\code{\link{enpy}()},
\code{\link{initest_options}()},
\code{\link{mstep_options}()},
\code{\link{pense_options}()},
\code{\link{pensem}()}
}
\concept{deprecated functions}
