% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{en_cd_options}
\alias{en_cd_options}
\title{Use Coordinate Descent to Solve Elastic Net Problems}
\usage{
en_cd_options(max_it = 1000, reset_it = 8)
}
\arguments{
\item{max_it}{maximum number of iterations.}

\item{reset_it}{number of iterations after which the residuals are
re-computed from scratch, to prevent numerical drifts from incremental
updates.}
}
\description{
Use Coordinate Descent to Solve Elastic Net Problems
}
\seealso{
Other EN algorithms: 
\code{\link{en_admm_options}()},
\code{\link{en_dal_options}()},
\code{\link{en_lars_options}()}
}
\concept{EN algorithms}
