% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{en_lars_options}
\alias{en_lars_options}
\title{Use the LARS Elastic Net Algorithm}
\usage{
en_lars_options()
}
\description{
Use the LARS Elastic Net Algorithm
}
\seealso{
Other EN algorithms: 
\code{\link{en_admm_options}()},
\code{\link{en_dal_options}()}
}
\concept{EN algorithms}
