% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.R
\name{enpy.control}
\alias{enpy.control}
\title{Control Parameters for ENPY}
\usage{
enpy.control(en.maxit = 50000, en.tol = 1e-08, en.centering = TRUE,
  mscale.maxit = 200, mscale.tol = 1e-08, mscale.rho.fun = c("bisquare",
  "huber", "gauss"))
}
\arguments{
\item{en.maxit}{Maximum number of iterations allowed for the elastic net algorithm.}

\item{en.tol}{Convergence threshold for elastic net.}

\item{en.centering}{Should the data be centered for the elastic net algorithm. Only those rows
with leading 1's will be considered for centering.}

\item{mscale.maxit}{Maximum number of iterations allowed for the m-scale algorithm.}

\item{mscale.tol}{Convergence threshold for the m-scale}

\item{mscale.rho.fun}{The rho function to use for the m-scale.}
}
\value{
A list with the given arguments.
}
\description{
Set control parameters for ENPY.
}

