% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{enpy}
\alias{enpy}
\title{Deprecated: ENPY Initial Estimates for EN S-Estimators}
\usage{
enpy(x, y, alpha, lambda, delta, cc, options, en_options)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge('deprecated')}
}
\details{
Superseded by \code{\link[=enpy_initial_estimates]{enpy_initial_estimates()}}.
}
\keyword{internal}
