% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpy.R
\name{enpy}
\alias{enpy}
\title{PY (Pena-Yohai) initial estimates for the EN S-estimator}
\usage{
enpy(X, y, alpha, lambda, options = initest_options(),
  en_options = en_options_aug_lars())
}
\arguments{
\item{X}{data matrix with predictors.}

\item{y}{response vector.}

\item{alpha, lambda}{EN penalty parameters (NOT adjusted for the number of
observations in \code{X}).}

\item{options}{additional options for the initial estimator. See
\code{\link{initest_options}} for details.}

\item{en_options}{additional options for the EN algorithm. See
\code{\link{en_options}} for details.}
}
\value{
\item{initCoef}{A numeric matrix with one initial coefficient per column}
        \item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
Computes the PY initial estimates for the EN S-estimator with different
strategies for computing the principal sensitivity components.
}
\details{
Two different methods to calculate the sensitivity components are implemented:
\describe{
     \item{\code{"rr"}}{Approximate the PSCs by using the residuals from the
         elastic net fit and the hat matrix from the ridge regression.
         This method only works if \code{alpha} < 1 or
         \code{ncol(X)} < \code{nrow(X)}.}
     \item{\code{"exact"}}{Calculate the PSCs from the difference between the
         residuals and leave-one-out residuals from elastic net.}
}
}
