% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enpy.R
\name{enpy}
\alias{enpy}
\title{PY (Peña-Yohai) initial estimates for EN}
\usage{
enpy(X, y, lambda1, lambda2, deltasc, cc.scale, psc.method = c("rr", "Qp",
  "Mn"), prosac, clean.method = c("threshold", "proportion"), C.res, prop,
  py.nit, en.tol, en.centering = TRUE)
}
\arguments{
\item{X}{The data matrix X}

\item{y}{The response vector}

\item{lambda1,lambda2}{The EN penalty parameters (adjusted for the number of observations
in \code{X})}

\item{deltasc,cc.scale}{Parameters for the M-equation of the scale. Tukey's bisquare rho function
is used internally.}

\item{psc.method}{The method to use for computing the principal sensitivity components.
See details.}

\item{prosac}{The proportion of observations to remove based on PSCs}

\item{clean.method}{How to clean the data based on large residuals.
If \code{"threshold"}, all observations with scaled residuals larger than
\code{C.res} will be removed.
If \code{"proportion"}, observations with the largest \code{prop} residuals
will be removed.}

\item{py.nit}{The maximum number of iterations to perform.}

\item{en.tol}{The relative tolerance for convergence.}

\item{en.centering}{Should rows with a leading 1 be centered in the elastic net algorithm.
     Default's to \code{TRUE}.}
}
\value{
\item{initCoef}{A numeric matrix with one initial coefficient per column}
        \item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
Computes the PY initial estimates for EN with different strategies for computing
the principal sensitivity components
}
\details{
Three different methods to calculate the sensitivity components are implemented:
\describe{
     \item{\code{"rr"}}{Approximate the PSCs by using the residuals from the elastic net fit
                        and the hat matrix from the ridge regression. This method only works
                        if `lambda` > 0 or `p` < `n`.}
     \item{\code{"Qp"}}{Not yet implemented}
     \item{\code{"Mn"}}{Not yet implemented}
}
}

