% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enpy.R
\name{enpy}
\alias{enpy}
\title{PY (Peña-Yohai) initial estimates for EN}
\usage{
enpy(X, y, lambda1, lambda2, deltaesc, cc.scale, psc.method = c("rr", "Mn",
  "Qp"), prosac, clean.method = c("threshold", "proportion"), C.res = NULL,
  prop = NULL, py.nit, en.tol, control = enpy.control())
}
\arguments{
\item{X}{The data matrix X.}

\item{y}{The response vector.}

\item{lambda1,lambda2}{The EN penalty parameters (NOT adjusted for the number of observations
in \code{X}).}

\item{deltaesc,cc.scale}{Parameters for the M-equation of the scale. If \code{cc.scale}
is missing or invalid, it will be chosen such that the expected value of the
rho function under the normal model is equal to \code{deltaesc}.
The default rho function (Tukey's bisquare) can be changed by
parameter \code{control}.}

\item{psc.method}{The method to use for computing the principal sensitivity components.
See details.}

\item{prosac}{The proportion of observations to remove based on PSCs.}

\item{clean.method}{How to clean the data based on large residuals.
If \code{"threshold"}, all observations with scaled residuals larger than
\code{C.res} will be removed.
If \code{"proportion"}, observations with the largest \code{prop} residuals
will be removed.}

\item{py.nit}{The maximum number of iterations to perform.}

\item{en.tol}{The relative tolerance for convergence.}

\item{control}{Optional further control parameters from \code{\link{enpy.control}}.}
}
\value{
\item{initCoef}{A numeric matrix with one initial coefficient per column}
        \item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
Computes the PY initial estimates for EN with different strategies for computing
the principal sensitivity components
}
\details{
Three different methods to calculate the sensitivity components are implemented:
\describe{
     \item{\code{"rr"}}{Approximate the PSCs by using the residuals from the elastic net fit
                        and the hat matrix from the ridge regression. This method only works
                        if \code{lambda2} > 0 or \code{ncol(X)} < \code{nrow(X)}.}
     \item{\code{"Mn"}}{Calculate the PSCs from the difference between the
                        residuals and leave-one-out residuals from elastic net.}
}
}

