% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{enpy}
\alias{enpy}
\title{Deprecated}
\usage{
enpy(x, y, alpha, lambda, delta, cc, options, en_options)
}
\arguments{
\item{x}{data matrix with predictors.}

\item{y}{response vector.}

\item{alpha, lambda}{EN penalty parameters (NOT adjusted for the number of
observations in \code{x}).}

\item{delta}{desired breakdown point of the resulting estimator.}

\item{cc}{tuning constant for the S-estimator. Default is to chosen based
on the breakdown point \code{delta}. Should never have to be changed.}

\item{options}{\strong{ignored.} Additional options for the initial estimator.}

\item{en_options}{\strong{ignored.} Additional options for the EN algorithm.}
}
\value{
\item{coeff}{A numeric matrix with one initial coefficient per column}
\item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge('deprecated')}

Compute initial estimates for EN S-estimates using ENPY.
Superseded by \code{\link[=enpy_initial_estimates]{enpy_initial_estimates()}}.
}
\section{Warning}{
Do not use this function in new code.
It may be removed from future versions of the package.
}

\seealso{
Other deprecated functions: 
\code{\link{deprecated_en_options}},
\code{\link{initest_options}()},
\code{\link{mstep_options}()},
\code{\link{pense_options}()},
\code{\link{pensem}()}
}
\concept{deprecated functions}
