% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initest.control.R
\name{initest.control}
\alias{initest.control}
\title{Creates the internal control list for PY initial estimators}
\usage{
initest.control(lambda1, lambda2, numIt, eps = 1e-06,
  residCleanMethod = c("proportion", "threshold"), residThreshold,
  residProportion, pscProportion, mscaleB = 0.5, mscaleCC = 1.54764,
  enpy.control)
}
\arguments{
\item{lambda1}{numeric >= 0}

\item{lambda2}{numeric >= 0}

\item{numIt}{integer > 0}

\item{eps}{numeric > 0}

\item{residCleanMethod}{character}

\item{residThreshold}{If \code{residCleanMethod = "threshold"} numeric > 0, otherwise not
referenced}

\item{residProportion}{If \code{residCleanMethod = "proportion"} numeric > 0, otherwise not
referenced}

\item{pscProportion}{numeric > 0}

\item{mscaleB}{numeric > 0}

\item{mscaleCC}{numeric > 0}

\item{enMaxIt}{integer > 0}

\item{enEPS}{numeric > 0}

\item{enCentering}{logical}

\item{mscaleMaxIt}{integer > 0}

\item{mscaleEPS}{numeric > 0}

\item{mscaleRhoFun}{character}
}
\description{
Computes the PY initial estimates for EN with approximated principal sensitivity components
by the ridge regression solution.
}

