% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{initest_options}
\alias{initest_options}
\title{Deprecated}
\usage{
initest_options(
  keep_solutions = 5,
  psc_method = c("exact", "rr"),
  maxit = 10,
  maxit_pense_refinement = 5,
  eps = 1e-06,
  psc_keep = 0.5,
  resid_keep_method = c("proportion", "threshold"),
  resid_keep_prop = 0.6,
  resid_keep_thresh = 2,
  mscale_eps = 1e-08,
  mscale_maxit = 200
)
}
\arguments{
\item{keep_solutions}{how many initial estimates should be kept to perform
full PENSE iterations?}

\item{psc_method}{The method to use for computing the principal sensitivity
components. See details for the possible choices.}

\item{maxit}{maximum number of refinement iterations.}

\item{maxit_pense_refinement}{\strong{ignored.} Maximum number of PENSE iterations to refine
initial estimator.}

\item{eps}{\strong{ignored.} Numeric tolerance for convergence.}

\item{psc_keep}{proportion of observations to keep based on the PSC scores.}

\item{resid_keep_method}{How to clean the data based on large residuals.
If \code{"proportion"}, observations with the smallest
\code{resid_keep_prop} residuals will be retained.
If \code{"threshold"}, all observations with scaled residuals smaller
than the threshold \code{resid_keep_thresh} will be retained.}

\item{resid_keep_prop, resid_keep_thresh}{proportion or threshold for
observations to keep based on their residual.}

\item{mscale_eps, mscale_maxit}{\strong{ignored.} Maximum number of iterations and numeric
tolerance for the M-scale.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge('deprecated')}

Options for computing initial estimates via ENPY.
Superseded by \code{\link[=enpy_options]{enpy_options()}}.
}
\section{Warning}{
Do not use this function in new code.
It may be removed from future versions of the package.
}

\seealso{
Other deprecated functions: 
\code{\link{deprecated_en_options}},
\code{\link{enpy}()},
\code{\link{mstep_options}()},
\code{\link{pense_options}()},
\code{\link{pensem}()}
}
\concept{deprecated functions}
