% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscale.R
\name{mscale}
\alias{mscale}
\title{Robust M-estimate of Scale}
\usage{
mscale(x, b = 0.5, rho = c("bisquare", "huber"), cc, eps = 1e-08,
  max.it = 200)
}
\arguments{
\item{x}{A numeric vector.}

\item{b}{The value of the M-estimation equation.}

\item{rho}{The rho function to use in the M-estimation equation.}

\item{cc}{Non-negative constant for the chosen rho function. If missing, it will be
chosen such that the expected value of the rho function under the normal model
is equal to \code{b}.}

\item{eps}{Threshold for convergence}

\item{max.it}{The maximum number of iterations}
}
\value{
Numeric vector of length one
}
\description{
Compute the M-estimate of scale with MAD as initial estimate.
}
\details{
This solves the M-estimation equation given by
\deqn{\sum_{i=1}^n \rho( x_i / s_n; cc ) = n b}

\code{NA} values in \code{x} are removed before calculating the
}
