% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{mstep}
\alias{mstep}
\title{Perform an M-step on the PENSE result}
\usage{
mstep(penseobj, complete.grid = TRUE, cv.k = 5L, nlambda = 30L,
  ncores = getOption("mc.cores", 1L), cl = NULL)
}
\arguments{
\item{penseobj}{an object returned from a call to \code{\link{pense}}.}

\item{complete.grid}{should the optimal lambda be chosen from a grid of values
or approximate it from the optimal lambda}

\item{cv.k}{perform k-fold CV to choose optimal lambda (only used if
\code{complete.grid = FALSE}).}

\item{nlambda}{the number of lambda values to try.}

\item{ncores, cl}{use this many cores or the supplied cluster for choosing the
optima lambda. See \code{\link{pense}} for more details.}
}
\description{
Performs an M-step on the S-estimator returned from \code{\link{pense}}.
}
