% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstep.R
\name{mstep}
\alias{mstep}
\title{Perform an M-step after the EN S-Estimator}
\usage{
mstep(penseobj, lambda, complete_grid = TRUE, cv_k = 5L, nlambda = 30L,
  ncores = getOption("mc.cores", 1L), cl = NULL,
  options = mstep_options(), X, y, cv_objective, lambda_min_ratio, en_options)
}
\arguments{
\item{penseobj}{an object returned from a call to \code{\link{pense}}.}

\item{lambda}{regularization of the initial S-estimate. Defaults to
\code{penseobj$lambda_opt}, the lambda with minimal CV error.}

\item{complete_grid}{should the optimal lambda be chosen from a grid of
values or approximate it from the optimal lambda?}

\item{cv_k}{perform k-fold CV to choose optimal lambda (only used if
\code{complete_grid = TRUE}).}

\item{nlambda}{the number of lambda values to try.}

\item{ncores, cl}{use this many cores or the supplied cluster for choosing the
optima lambda. See \code{\link{pense}} for more details.}

\item{options}{additional options for the M-step.}

\item{X, y, cv_objective, lambda_min_ratio, en_options}{override arguments
provided to the original call to \code{\link{pense}}.}

\item{adjust_bdp}{should the breakdown point be adjusted based on the
effective degrees of freedom?}
}
\description{
Performs an M-step on the S-estimator returned from \code{\link{pense}}.
}
