% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-options.R
\name{mstep_options}
\alias{mstep_options}
\title{Additional Options for the Penalized EN M-estimator}
\usage{
mstep_options(cc = 3.44, maxit = 1000, eps = 1e-06)
}
\arguments{
\item{cc}{tuning constant for the M-estimator.}

\item{maxit}{maximum number of iterations allowed.}

\item{eps}{numeric tolerance for convergence.}
}
\value{
a checked options list.
}
\description{
Additional Options for the Penalized EN M-estimator
}
\seealso{
Other specifying additional options: \code{\link{en_options_aug_lars}},
  \code{\link{initest_options}},
  \code{\link{pense_options}}
}
