% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{mstep_options}
\alias{mstep_options}
\title{Deprecated}
\usage{
mstep_options(
  cc = 3.44,
  maxit = 1000,
  eps = 1e-06,
  adjust_bdp = FALSE,
  verbosity = 0,
  en_correction = TRUE
)
}
\arguments{
\item{cc}{\strong{ignored.} Tuning constant for the M-estimator.}

\item{maxit}{maximum number of iterations allowed.}

\item{eps}{\strong{ignored.} Numeric tolerance for convergence.}

\item{adjust_bdp}{\strong{ignored.} Should the breakdown point be adjusted based on the
effective degrees of freedom?}

\item{verbosity}{\strong{ignored.} Verbosity of the algorithm.}

\item{en_correction}{\strong{ignored.} Should the corrected EN estimator be used to choose
the optimal lambda with CV.
If \code{TRUE}, as by default, the estimator is "bias corrected".}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge('deprecated')}

Additional options for computing penalized EN MM-estimates.
Superseded by \code{\link[=mm_algorithm_options]{mm_algorithm_options()}} and options supplied directly to \code{\link[=pensem_cv]{pensem_cv()}}.
}
\section{Warning}{
Do not use this function in new code.
It may be removed from future versions of the package.
}

\seealso{
Other deprecated functions: 
\code{\link{deprecated_en_options}},
\code{\link{enpy}()},
\code{\link{initest_options}()},
\code{\link{pense_options}()},
\code{\link{pensem}()}
}
\concept{deprecated functions}
