% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.R
\name{pense.control}
\alias{pense.control}
\title{Control Parameters for PENSE}
\usage{
\method{pense}{control}(pense.maxit = 500, pense.tol = 1e-06, pense.en.tol,
  pense.en.maxit = 5e+05, cv.objective = scaleTau2, init.tol = 1e-06,
  init.maxit = 10, init.resid.clean.method = c("proportion", "threshold"),
  init.resid.proportion = 0.4, init.resid.threshold = 2,
  init.psc.method = c("auto", "rr", "Mn"), init.psc.proportion = 0.8,
  init.csteps = 10, init.nkeep = 5, init.en.tol,
  init.en.maxit = pense.en.maxit, mscale.cc, mscale.delta = 0.5,
  mscale.maxit = 200, mscale.tol = 1e-08, mscale.rho.fun = c("bisquare",
  "huber"), en.algorithm = c("augmented-lars", "coordinate-descent",
  "augmented-lars-gram", "augmented-lars-nogram"))
}
\arguments{
\item{pense.maxit}{maximum number of iterations for PENSE.}

\item{pense.tol}{convergence tolerance for PENSE.}

\item{pense.en.tol}{The relative tolerance for convergence for gradient-descent (default 1e-9) or
the threshold for treating numbers as 0 in LARS (default .Machine$double.eps) in the
PENSE step.}

\item{pense.en.maxit}{maximum number of iterations for each elastic net fit in PENSE}

\item{cv.objective}{a user-supplied function taking the cross-validated residuals
         to choose lambda. Default is to choose the lambda the minimizes the
         \link[robustbase:scaleTau2]{tau-scale}.}

\item{init.tol}{convergence tolerance for initial estimates.}

\item{init.maxit}{maximum number of iterations for initial estimates.}

\item{init.resid.clean.method}{how to clean the data based on large residuals.
If \code{"threshold"}, all observations with scaled residuals larger than
\code{init.resid.threshold} will be removed (\code{init.resid.threshold}
corresponds to the constant \eqn{C_1} from equation (21) in Pena & Yohai (1999).
If \code{"proportion"}, observations with the largest
\code{init.resid.proportion} residuals will be removed.}

\item{init.resid.proportion,init.resid.threshold}{see details of argument
\code{init.resid.clean.method}}

\item{init.psc.method}{how to calculate the principal sensitivity components. If \code{"auto"},
\code{rr} is used when there are enough observations, otherwise \code{Mn} is
selected. See \code{\link{enpy}} for details.}

\item{init.psc.proportion}{the proportion of observations to remove based on the score on
the principal sensitivity components.}

\item{init.csteps}{the number of PENSE iterations (concentration steps) to perform on
the initial estimates.}

\item{init.nkeep}{how many initial estimates should be kept to perform full PENSE iterations?}

\item{init.en.maxit}{maximum number of iterations for each elastic net fit for initial estimates.}

\item{mscale.cc}{tuning parameter for the M-estimate of scale. If missing or lte 0,
chosen such that the expected value under the normal model of the rho function
equals to \code{mscale.delta}.}

\item{mscale.delta}{determines the breakdown point of the S-estimator.}

\item{mscale.maxit}{maximum number of iterations to calculate the M-estimate of scale.}

\item{mscale.tol}{convergence tolerance for calculating the M-estimate of scale.}

\item{mscale.rho.fun}{string specifying the rho function of the S-estimator.}

\item{en.algorithm}{algorithm to use to compute the (weighted) elastic net solutions for the
         inital estimates and PENSE.}

\item{pense.en.tol}{The relative tolerance for convergence for gradient-descent (default 1e-8) or
the threshold for treating numbers as 0 in LARS (default .Machine$double.eps) for
the initial estimate.}
}
\value{
A list with the given arguments, checked for errors.
}
\description{
Set control parameters for PENSE.
}
\references{
Pena, D., & Yohai, V.. (1999). A Fast Procedure for Outlier Diagnostics in Large
Regression Problems. \emph{Journal of the American Statistical Association}, 94(446),
434–445. \url{http://doi.org/10.2307/2670164}
}

