% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{pense_options}
\alias{pense_options}
\title{Deprecated}
\usage{
pense_options(
  delta = 0.25,
  maxit = 1000,
  eps = 1e-06,
  mscale_eps = 1e-08,
  mscale_maxit = 200,
  verbosity = 0,
  cc = NULL,
  en_correction = TRUE
)
}
\arguments{
\item{delta}{desired breakdown point of the resulting estimator.}

\item{maxit}{maximum number of iterations allowed.}

\item{eps}{numeric tolerance for convergence.}

\item{mscale_eps, mscale_maxit}{maximum number of iterations and numeric
tolerance for the M-scale.}

\item{verbosity}{\strong{ignored.} Verbosity of the algorithm.}

\item{cc}{\strong{ignored.} Tuning constant for the S-estimator. Default is to chosen based
on the breakdown point \code{delta}. Should never have to be changed.}

\item{en_correction}{\strong{ignored.} Should the corrected EN estimator be used to choose
the optimal lambda with CV.
If \code{TRUE}, as by default, the estimator is "bias corrected".}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge('deprecated')}

Additional options for computing penalized EN S-estimates.
Superseded by \code{\link[=mm_algorithm_options]{mm_algorithm_options()}} and options supplied directly to \code{\link[=pense]{pense()}}.
}
\section{Warning}{
Do not use this function in new code.
It may be removed from future versions of the package.
}

\seealso{
Other deprecated functions: 
\code{\link{deprecated_en_options}},
\code{\link{enpy}()},
\code{\link{initest_options}()},
\code{\link{mstep_options}()},
\code{\link{pensem}()}
}
\concept{deprecated functions}
