% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensem_regression.R
\name{pensem}
\alias{pensem}
\title{Deprecated Alias of pensem_cv}
\usage{
pensem(x, ...)
}
\arguments{
\item{x}{either a numeric matrix of predictor values, or a cross-validated PENSE fit from \code{\link[=pense_cv]{pense_cv()}}.}

\item{...}{ignored. See the section on deprecated parameters below.}
}
\description{
\code{pensem()} is a deprecated alias for \code{\link[=pensem_cv]{pensem_cv()}}.
}
\seealso{
Other deprecated functions: 
\code{\link{deprecated_en_options}},
\code{\link{enpy}()},
\code{\link{initest_options}()},
\code{\link{mstep_options}()},
\code{\link{pense_options}()}
}
\concept{deprecated functions}
