% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.pense.R
\name{predict.elnetfit}
\alias{predict.elnetfit}
\title{Predict Method for Penalized Elastic-Net Estimator}
\usage{
\method{predict}{elnetfit}(object, newdata, lambda, exact = FALSE, ...)
}
\arguments{
\item{object}{an object of type \code{elnetfit} to use for prediction.}

\item{newdata}{an optional design matrix}

\item{lambda}{the value of the penalty parameter. Default is to use the
optimal lambda \code{lambda_opt}.}

\item{exact}{if the lambda is not part of the lambda grid, should the
estimates be obtained by linear interpolation between the nearest
lambda values (default) or computed exactly.}

\item{...}{currently ignored.}
}
\value{
a numeric vector of predicted values for the given lambda.
}
\description{
Predict Method for Penalized Elastic-Net Estimator
}
