% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prinsens.R
\name{prinsens}
\alias{prinsens}
\title{Principal Sensitivity Components}
\usage{
prinsens(X, y, method = c("ols", "en"), intercept = TRUE, alpha, lambda,
  en_options = en_options_aug_lars())
}
\arguments{
\item{X}{data matrix with predictors}

\item{y}{response vector}

\item{method}{use ordinary least squares (\code{"ols"}) or elastic net
(\code{"en"}) to compute the PSCs.}

\item{intercept}{Should an intercept be added or not.}

\item{alpha, lambda}{The values for the parameters controlling the
penalization for elastic net.}

\item{en_options}{additional options for the EN algorithm. See
\code{\link{en_options}} for details.}
}
\value{
A numeric matrix with as many rows as \code{X} and as many columns as
     PSCs found (at most the number of columns in \code{X} plus one for the
     intercept). Each column is a PSC.
}
\description{
Compute the principal sensitivity components (PSC) for regression.
}
\references{
Pena, D., and Yohai, V.J. (1999).
    A Fast Procedure for Outlier Diagnostics in Large Regression Problems.
    \emph{Journal of the American Statistical Association}, \bold{94}(446),
    434-445. \url{http://doi.org/10.2307/2670164}
}
