% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prinsens.R
\name{prinsens}
\alias{prinsens}
\title{Principal Sensitivity Components}
\usage{
prinsens(X, y, method = c("ols", "en"), intercept = TRUE, alpha, lambda,
  maxit = 50000, eps = 1e-08, centering = TRUE)
}
\arguments{
\item{X}{The data matrix X}

\item{y}{The response vector}

\item{method}{Use ordinary least squares (\code{"ols"}) or elastic net (\code{"en"}) to compute
the PSCs.}

\item{intercept}{Should an intercept be added or not.}

\item{alpha,lambda}{The values for the parameters controlling the penalization for elastic net.}

\item{maxit}{The maximum number of iterations for elastic net.}

\item{eps}{The relative tolerance for convergence for elastic net.}

\item{centering}{Should the rows be centered first for elastic net.}
}
\value{
A numeric matrix with as many rows as \code{X} and as many columns as
     PSCs found (at most the number of columns in \code{X} plus one for the intercept).
     Each column is a PSC.
}
\description{
Compute the principal sensitiviy components (PSC) for regression.
}
\references{
Pena, D., & Yohai, V.. (1999). A Fast Procedure for Outlier Diagnostics in Large
Regression Problems. \emph{Journal of the American Statistical Association}, 94(446),
434–445. \url{http://doi.org/10.2307/2670164}
}

