% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pyinit.R
\name{pyinit}
\alias{pyinit}
\title{PY (Peña-Yohai) initial estimates for S-estimates of regression}
\usage{
pyinit(X, y, deltasc, cc.scale, prosac, clean.method = c("threshold",
  "proportion"), C.res, prop, py.nit, en.tol)
}
\arguments{
\item{X}{The data matrix X}

\item{y}{The response vector}

\item{deltasc,cc.scale}{Parameters for the M-equation of the scale. Tukey's bisquare rho function
is used internally.}

\item{prosac}{The proportion of observations to remove based on PSCs}

\item{clean.method}{How to clean the data based on large residuals.
If \code{"threshold"}, all observations with scaled residuals larger than
\code{C.res} will be removed.
If \code{"proportion"}, observations with the largest \code{prop} residuals
will be removed.}

\item{py.nit}{The maximum number of iterations to perform.}

\item{en.tol}{The relative tolerance for convergence.}
}
\value{
\item{initCoef}{A numeric matrix with one initial coefficient per column}
        \item{objF}{A vector of values of the objective function for the respective coefficient}
}
\description{
Computes the PY initial estimates for S-estimates of regression.
}
\references{
Pena, D., & Yohai, V.. (1999). A Fast Procedure for Outlier Diagnostics in Large
Regression Problems. \emph{Journal of the American Statistical Association}, 94(446),
434–445. \url{http://doi.org/10.2307/2670164}
}

