% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rho_function}
\alias{rho_function}
\title{List Available Rho Functions}
\usage{
rho_function(rho)
}
\arguments{
\item{rho}{the name of the \eqn{\rho} function to check for existence.}
}
\value{
if \code{rho} is missing returns a vector of supported \eqn{\rho} function names, otherwise
the internal integer representation of the \eqn{\rho} function.
}
\description{
List Available Rho Functions
}
\seealso{
Other miscellaneous functions: 
\code{\link{consistency_const}()}
}
\concept{miscellaneous functions}
